requirements.txt
# Core API
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0

# Database
motor==3.4.0
pymongo==4.7.2
redis[hiredis]==5.0.4

# Task Queue
celery[redis]==5.4.0
kombu==5.3.7

# Auth
firebase-admin==6.5.0
PyJWT==2.8.0

# Security / Encryption
cryptography==42.0.8
python-magic==0.4.27

# HTTP client
httpx==0.27.0

# Media processing
Pillow==10.3.0
pdf2image==1.17.0

# Validation
pydantic[email]==2.7.1
pydantic-settings==2.3.0

# Structured logging
structlog==24.2.0

# Payments
stripe==9.9.0
razorpay==1.4.0

# Email
resend==0.8.0

# Monitoring
sentry-sdk[fastapi,celery]==2.4.0
prometheus-fastapi-instrumentator==7.0.0

# NTP (beat clock check)
ntplib==0.4.0

# Rate limiting
slowapi==0.1.9

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0
respx==0.21.1
testcontainers[mongodb,redis]==4.5.1
celery[pytest]==5.4.0

# Code quality / security
bandit==1.7.9
semgrep==1.77.0
pip-audit==2.7.3
